# api.yml
openapi: 3.0.1
info:
  title: ExtranetAPI
  version: '1.0'
paths:
...
  /buisness-costs:
    get:
      tags:
        - BuisnessCost
      summary: Получить все бизнесовые издержки
      responses:
        '200':
          description: Получить все издержки
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BuisnessCost'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BuisnessCost'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BuisnessCost'
    post:
      tags:
        - BuisnessCost
      summary: Создать бизнесовую издержку
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BuisnessCost'
          text/json:
            schema:
              $ref: '#/components/schemas/BuisnessCost'
          application/*+json:
            schema:
              $ref: '#/components/schemas/BuisnessCost'
        required: true
      responses:
        '200':
          description: Создать издержку
          content:
            text/plain:
              schema:
                type: integer
                format: int32
            application/json:
              schema:
                type: integer
                format: int32
            text/json:
              schema:
                type: integer
                format: int32
  /buisness-costs/{id}:
    delete:
      tags:
        - BuisnessCost
      summary: Удалить бизнесовую издержку
      parameters:
        - name: id
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Удалить издержку
          content:
            text/plain:
              schema:
                type: integer
                format: int32
            application/json:
              schema:
                type: integer
                format: int32
            text/json:
              schema:
                type: integer
                format: int32
...
  /costs/{costId}:
    get:
      tags:
        - Cost
      summary: Получить издержку по идентификатору
      parameters:
        - name: costId
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Получить издержку по идентификатору
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/Cost'
            application/json:
              schema:
                $ref: '#/components/schemas/Cost'
            text/json:
              schema:
                $ref: '#/components/schemas/Cost'
  /costs:
    get:
      tags:
        - Cost
      summary: Получить все издержки
      responses:
        '200':
          description: Получить все издержки
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cost'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cost'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cost'
    post:
      tags:
        - Cost
      summary: Создать издержку
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Cost'
          text/json:
            schema:
              $ref: '#/components/schemas/Cost'
          application/*+json:
            schema:
              $ref: '#/components/schemas/Cost'
        required: true
      responses:
        '200':
          description: Создать издержку
          content:
            text/plain:
              schema:
                type: integer
                format: int32
            application/json:
              schema:
                type: integer
                format: int32
            text/json:
              schema:
                type: integer
                format: int32
  /projects/{projectId}:
    get:
      tags:
        - Project
      summary: Получить проект по идентификатору
      parameters:
        - name: projectId
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Получить проект по идентификатору
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProjectDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ProjectDto'
    delete:
      tags:
        - Project
      summary: Удалить проект
      parameters:
        - name: projectId
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Удалить проект
          content:
            text/plain:
              schema:
                type: integer
                format: int32
            application/json:
              schema:
                type: integer
                format: int32
            text/json:
              schema:
                type: integer
                format: int32
  /projects:
    get:
      tags:
        - Project
      summary: Получить все проекты
      responses:
        '200':
          description: Получить все проекты
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectDto'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectDto'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectDto'
    post:
      tags:
        - Project
      summary: Создать проект
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectDto'
          text/json:
            schema:
              $ref: '#/components/schemas/ProjectDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/ProjectDto'
        required: true
      responses:
        '200':
          description: Создать проект
          content:
            text/plain:
              schema:
                type: integer
                format: int32
            application/json:
              schema:
                type: integer
                format: int32
            text/json:
              schema:
                type: integer
                format: int32
...
  /projects/{projectId}/deadline:
    post:
      tags:
        - Project
      summary: Установить ожидаемую дату завершения выполнения
      parameters:
        - name: projectId
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int32
        - name: deadLine
          in: query
          description: ''
          required: true
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Установить ожидаемую дату завершения выполнения
          content:
            text/plain:
              schema:
                type: integer
                format: int32
            application/json:
              schema:
                type: integer
                format: int32
            text/json:
              schema:
                type: integer
                format: int32
...
  /project-deadline-settings/project-deadline-settings-updating:
    post:
      tags:
        - ProjectDeadline
      summary: Обновить информацию по дедлайну
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectDeadlineSettings'
          text/json:
            schema:
              $ref: '#/components/schemas/ProjectDeadlineSettings'
          application/*+json:
            schema:
              $ref: '#/components/schemas/ProjectDeadlineSettings'
        required: true
      responses:
        '200':
          description: Обновить основную информацию по проекту
          content:
            text/plain:
              schema:
                type: integer
                format: int32
            application/json:
              schema:
                type: integer
                format: int32
            text/json:
              schema:
                type: integer
                format: int32
...
components:
  schemas:
...
    BuisnessCost:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
          nullable: true
        amount:
          type: number
          format: double
        date:
          type: string
          format: date-time
      additionalProperties: false
...
    Cost:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
          nullable: true
      additionalProperties: false
...
    ProjectDeadlineSettings:
      type: object
      properties:
        id:
          type: integer
          format: int32
        daysForDeadlineYellow:
          type: integer
          format: int32
        daysForDeadlineRed:
          type: integer
          format: int32
        defaultProjectDurationDays:
          type: integer
          format: int32
      additionalProperties: false
    ProjectDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
        projectType:
          type: string
          nullable: true
        contractNumber:
          type: string
          nullable: true
        dateOfStart:
          type: string
          format: date-time
          nullable: true
        dateOfApplication:
          type: string
          format: date-time
        deadLine:
          type: string
          format: date-time
          nullable: true
        endDate:
          type: string
          format: date-time
          nullable: true
        clientId:
          type: integer
          format: int32
        address:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        isCompleted:
          type: boolean
      additionalProperties: false
...